pyproject.toml
[tool.poetry]
name = "dorado-scheduling"
version = "0.1.0"
description = "Dorado observation planning and scheduling simulations"
authors = ["Leo Singer <leo.singer@ligo.org>"]
license = "NOSA"
packages = [
    {include = "dorado"}]
include = [
    "dorado/scheduling/data/*.txt",
    "dorado/scheduling/data/*.md"]

[tool.poetry.dependencies]
python = "^3.7"
skyfield = "*"
astropy = "*"
astroplan = "*"
mip = "*"
"ligo.skymap" = "^0.5.0"
healpy = "*"
zstandard = "*"

[tool.poetry.dev-dependencies]
flake8 = "*"

[tool.poetry.scripts]
dorado-scheduling = "dorado.scheduling.scripts.main:main"
dorado-scheduling-prepare = "dorado.scheduling.scripts.prepare:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
